DSN-028 (phase 5): extract composition root to internal/app
Fifth phase of the DSN-028 6-phase package restructure. The body of
cmd/main.go (~520 lines) extracts into internal/app/server.go as an
app.Server with a New/NewWithDeps/Run/Cleanup API. cmd/server/main.go
is now a ~50-line entry point: load config, signal handle, hand off
to app.Server.

Moves
- cmd/main.go body                  -> internal/app/server.go
  (Deps struct, Server struct, New, NewWithDeps, Run, Cleanup,
  ConfigLogging, PrintLogHeader, buildDeps + all build* helpers,
  configureSigner, startKafka, idempotentHandler, shutdownHTTP,
  resolveShutdownTimeout)
- cmd/main.go entry                 -> cmd/server/main.go (~50 lines)
- api/api.go (ConfigureRouter etc.) -> internal/app/routes.go
- api/api_test.go                   -> internal/app/routes_test.go
- api/health.go + test              -> internal/app/health{,_test}.go
- api/environmentapi*.go,
  api/environmentdto.go             -> internal/app/env{,_test,_dto}.go
- api/docs*.go                      -> internal/app/docs{,_test}.go
- api/authapi_test.go               -> internal/app/auth_integration_test.go
- api/auth_metrics_test.go          -> internal/app/auth_metrics_integration_test.go
- api/correlation_test.go           -> internal/app/correlation_integration_test.go
  (was temporarily moved to httpx in Phase 4; correlated to app
  router setup, belongs here)
- api/util*.go (Scrub)              -> internal/platform/httpx/scrub{,_test}.go
- api/openapi.yaml, openapi.json    -> internal/app/openapi.yaml, .json
  (artifacts colocate with the package that embeds them via go:embed)
- cmd/shutdown_test.go              -> internal/app/shutdown_test.go
  (it tests shutdownHTTP which now lives in app)
- cmd/integration_test.go           -> cmd/server/integration_test.go
- cmd/main.go                       DELETED

Composition root testability
internal/app/server_test.go exercises the acceptance criterion: a
unit test constructs app.Server via NewWithDeps using a fake
inventory service and the existing user.MockUserService, hits /live
via httptest.NewServer(srv.Handler()), and asserts the router is
wired without opening any external resource (no Postgres, no Redis,
no Kafka, no AMQP). NewWithDeps is the seam.

Deps struct shape
Plain struct of dependencies, as agreed in the approaches doc:
- DB, Redis (external connections)
- InventorySvc, UserService (domain services; InventorySvc satisfies
  both inventory.InventoryService and inventory.ReservationService)
- Signer (auth)
- CatalogClient (outbound)
- ReadinessDeps map (for /ready)
- IdempotencyMw, AuthRateLimitMw (HTTP middleware)
- TracingShutdown, KafkaCleanup (shutdown closures)

State after this phase
- api/ contains ONLY api/client/v1/ (the generated v1 Go client).
  No api package source files; no api directory in the Go import
  graph. cmd/demo and config still reference api/client/v1 by its
  module path.
- core/, db/, events/, idempotency/, kafka/, queue/, testutil/
  remain non-existent (Phase 4 work).
- cmd/main.go gone; cmd/server/main.go is the new entry point.

Build path updates
- Dockerfile: 'go build ./cmd' -> 'go build ./cmd/server';
  migrations COPY path updated to
  internal/platform/persistence/migrations
- Makefile: swag -g cmd/main.go -> cmd/server/main.go;
  swag output and oapi-codegen input -> internal/app/openapi.yaml
- config.go default db.migrationFolder: db/migrations ->
  internal/platform/persistence/migrations (was broken since
  Phase 4 but only surfaces in Docker / integration test paths)
- config_test.yml: matching migrationFolder path update

Schema names preserved
@name EnvResponse annotation keeps the short OpenAPI schema name.

Acceptance criteria addressed
- cmd/main.go body extracted into internal/app/server.go as
  app.Server with New(ctx, cfg) and (*Server) Run(ctx) error
- Router construction moves from api/api.go to internal/app/routes.go
- cmd/server/main.go is ~50 lines: load config, build app.Server,
  Run, exit
- cmd/main.go deleted
- cmd/integration_test.go moves to cmd/server/
- app.Server is constructible in a unit test with fakes for
  driven ports
- go build ./..., make verify pass

Diff size: 28 files changed, 643 insertions / 504 deletions with
--find-renames. ~140 net LOC; the gross is from cmd/main.go's body
being split across Deps, Server, build helpers, and lifecycle
methods which git can't fully detect as a single rename.
