+++ config/config.go
@@ -615,7 +615,7 @@ func setupDefaults(config *Config) {
 	config.Db.Name = StringConfig{Value: "micro-ex-db", Default: "micro-ex-db", Description: "The name of the database to connect to."}
 	config.Db.Host = StringConfig{Value: "5432", Default: "5432", Description: "Port of the database."}
 	config.Db.Migrate = BoolConfig{Value: true, Default: true, Description: "Whether or not database migrations should be executed on startup."}
-	config.Db.MigrationFolder = StringConfig{Value: "db/migrations", Default: "db/migrations", Description: "Location of migration files to be executed on startup."}
+	config.Db.MigrationFolder = StringConfig{Value: "internal/platform/persistence/migrations", Default: "internal/platform/persistence/migrations", Description: "Location of migration files to be executed on startup."}
 	config.Db.Clean = BoolConfig{Value: false, Default: false, Description: "WARNING: THIS WILL DELETE ALL DATA FROM THE DB. Used only during migration. If clean is true, all 'down' migrations are executed."}
 	config.Db.User = StringConfig{Value: "", Default: "", Description: "User the application will use to connect to the database. Supply via GME_DB_USER."}
 	config.Db.Pass = StringConfig{Value: "", Default: "", Description: "Password the application will use for connecting to the database. Supply via GME_DB_PASS."}
