+++ config_test.yml
@@ -15,7 +15,7 @@ db:
   port: 5432
   # user / pass: integration tests must supply GME_DB_USER, GME_DB_PASS.
   migrate: true
-  migrationFolder: ../db/migrations
+  migrationFolder: ../internal/platform/persistence/migrations
   clean: true
   logLevel: warn
   pool:
