+++ internal/app/docs.go
@@ -1,4 +1,4 @@
-package api
+package app
 
 import (
 	_ "embed"
