+++ internal/app/env.go
@@ -1,4 +1,4 @@
-package api
+package app
 
 import (
 	"net/http"
