+++ internal/app/openapi.yaml
@@ -32,6 +32,43 @@ components:
         requestID:
           type: string
       type: object
+    EnvResponse:
+      properties:
+        appName:
+          $ref: '#/components/schemas/config.StringConfig'
+        appVersion:
+          $ref: '#/components/schemas/config.StringConfig'
+        buildTime:
+          $ref: '#/components/schemas/config.StringConfig'
+        catalog:
+          $ref: '#/components/schemas/config.CatalogConfig'
+        config:
+          $ref: '#/components/schemas/config.ConfigSource'
+        db:
+          $ref: '#/components/schemas/config.DbConfig'
+        docs:
+          $ref: '#/components/schemas/config.DocsConfig'
+        idempotency:
+          $ref: '#/components/schemas/config.IdempotencyConfig'
+        kafka:
+          $ref: '#/components/schemas/config.KafkaConfig'
+        log:
+          $ref: '#/components/schemas/config.LogConfig'
+        port:
+          $ref: '#/components/schemas/config.StringConfig'
+        profile:
+          $ref: '#/components/schemas/config.StringConfig'
+        rabbitmq:
+          $ref: '#/components/schemas/config.QueueConfig'
+        rateLimit:
+          $ref: '#/components/schemas/config.RateLimitConfig'
+        redis:
+          $ref: '#/components/schemas/config.RedisConfig'
+        revision:
+          $ref: '#/components/schemas/config.StringConfig'
+        sha1Version:
+          $ref: '#/components/schemas/config.StringConfig'
+      type: object
     FieldProblem:
       properties:
         detail:
@@ -120,43 +157,6 @@ components:
           type: integer
         token_type:
           type: string
-      type: object
-    api.EnvResponse:
-      properties:
-        appName:
-          $ref: '#/components/schemas/config.StringConfig'
-        appVersion:
-          $ref: '#/components/schemas/config.StringConfig'
-        buildTime:
-          $ref: '#/components/schemas/config.StringConfig'
-        catalog:
-          $ref: '#/components/schemas/config.CatalogConfig'
-        config:
-          $ref: '#/components/schemas/config.ConfigSource'
-        db:
-          $ref: '#/components/schemas/config.DbConfig'
-        docs:
-          $ref: '#/components/schemas/config.DocsConfig'
-        idempotency:
-          $ref: '#/components/schemas/config.IdempotencyConfig'
-        kafka:
-          $ref: '#/components/schemas/config.KafkaConfig'
-        log:
-          $ref: '#/components/schemas/config.LogConfig'
-        port:
-          $ref: '#/components/schemas/config.StringConfig'
-        profile:
-          $ref: '#/components/schemas/config.StringConfig'
-        rabbitmq:
-          $ref: '#/components/schemas/config.QueueConfig'
-        rateLimit:
-          $ref: '#/components/schemas/config.RateLimitConfig'
-        redis:
-          $ref: '#/components/schemas/config.RedisConfig'
-        revision:
-          $ref: '#/components/schemas/config.StringConfig'
-        sha1Version:
-          $ref: '#/components/schemas/config.StringConfig'
       type: object
     config.BoolConfig:
       properties:
@@ -412,7 +412,7 @@ paths:
           content:
             application/json:
               schema:
-                $ref: '#/components/schemas/api.EnvResponse'
+                $ref: '#/components/schemas/EnvResponse'
           description: OK
         "401":
           content:
